Simple scrapping test
We navigate to horariodebuses page, send the form with the params received in the controller and return the HTML.
We are not really scrapping the data as the HTML structure sucks and it will take time to do a proper scrapping. This is just a test that shows that even though painful, it is possible to scrap the data.